api, backtest, lib, downloader, tqwebhelper, datetime: 将 api 中时间处理函数转移到单独的函数中，避免循环引用
