+++ tqsdk/api.py
@@ -1934,33 +1934,6 @@ class TqApi(object):
         return uuid.UUID(int=TqApi.RD.getrandbits(128)).hex
 
     @staticmethod
-    def _get_trading_day_start_time(trading_day):
-        """给定交易日, 获得交易日起始时间"""
-        begin_mark = 631123200000000000  # 1990-01-01
-        start_time = trading_day - 21600000000000  # 6小时
-        week_day = (start_time - begin_mark) // 86400000000000 % 7
-        if week_day >= 5:
-            start_time -= 86400000000000 * (week_day - 4)
-        return start_time
-
-    @staticmethod
-    def _get_trading_day_end_time(trading_day):
-        """给定交易日, 获得交易日结束时间"""
-        return trading_day + 64799999999999  # 18小时
-
-    @staticmethod
-    def _get_trading_day_from_timestamp(timestamp):
-        """给定时间, 获得其所属的交易日"""
-        begin_mark = 631123200000000000  # 1990-01-01
-        days = (timestamp - begin_mark) // 86400000000000  # 自 1990-01-01 以来的天数
-        if (timestamp - begin_mark) % 86400000000000 >= 64800000000000:  # 大于18点, 天数+1
-            days += 1
-        week_day = days % 7
-        if week_day >= 5:  # 如果是周末则移到星期一
-            days += 7 - week_day
-        return begin_mark + days * 86400000000000
-
-    @staticmethod
     def _deep_copy_dict(source, dest):
         for key, value in source.items():
             if isinstance(value, Entity):
