+++ tqsdk/backtest.py
@@ -8,6 +8,7 @@ import requests
 import asyncio
 from typing import Union
 from datetime import date, datetime
+from tqsdk.datetime import _get_trading_day_start_time, _get_trading_day_end_time
 from tqsdk.exceptions import BacktestFinished
 from tqsdk.objs import Entity
 import tqsdk.api
@@ -55,15 +56,13 @@ class TqBacktest(object):
         if isinstance(start_dt, datetime):
             self._start_dt = int(start_dt.timestamp() * 1e9)
         elif isinstance(start_dt, date):
-            self._start_dt = tqsdk.api.TqApi._get_trading_day_start_time(
-                int(datetime(start_dt.year, start_dt.month, start_dt.day).timestamp()) * 1000000000)
+            self._start_dt = _get_trading_day_start_time(int(datetime(start_dt.year, start_dt.month, start_dt.day).timestamp()) * 1000000000)
         else:
             raise Exception("回测起始时间(start_dt)类型 %s 错误, 请检查 start_dt 数据类型是否填写正确" % (type(start_dt)))
         if isinstance(end_dt, datetime):
             self._end_dt = int(end_dt.timestamp() * 1e9)
         elif isinstance(end_dt, date):
-            self._end_dt = tqsdk.api.TqApi._get_trading_day_end_time(
-                int(datetime(end_dt.year, end_dt.month, end_dt.day).timestamp()) * 1000000000)
+            self._end_dt = _get_trading_day_end_time(int(datetime(end_dt.year, end_dt.month, end_dt.day).timestamp()) * 1000000000)
         else:
             raise Exception("回测结束时间(end_dt)类型 %s 错误, 请检查 end_dt 数据类型是否填写正确" % (type(end_dt)))
         self._current_dt = self._start_dt
@@ -364,8 +363,7 @@ class TqBacktest(object):
                                 }
                             }
                             timestamp = item[
-                                "datetime"] if dur < 86400000000000 else tqsdk.api.TqApi._get_trading_day_start_time(
-                                item["datetime"])
+                                "datetime"] if dur < 86400000000000 else _get_trading_day_start_time(item["datetime"])
                             if timestamp > self._end_dt:  # 超过结束时间
                                 return
                             yield timestamp, diff, None  # K线刚生成时的数据都为开盘价
@@ -381,8 +379,7 @@ class TqBacktest(object):
                                 }
                             }
                             timestamp = item[
-                                            "datetime"] + dur - 1000 if dur < 86400000000000 else tqsdk.api.TqApi._get_trading_day_end_time(
-                                item["datetime"])
+                                            "datetime"] + dur - 1000 if dur < 86400000000000 else _get_trading_day_end_time(item["datetime"])
                             if timestamp > self._end_dt:  # 超过结束时间
                                 return
                             yield timestamp, diff, self._get_quotes_from_kline(self._data["quotes"][ins], timestamp,
