+++ tqsdk/tqwebhelper.py
@@ -15,6 +15,7 @@ from datetime import datetime
 from aiohttp import web
 import socket
 import tqsdk
+from tqsdk.datetime import _get_trading_day_start_time
 
 class TqWebHelper(object):
 
@@ -221,7 +222,7 @@ class TqWebHelper(object):
             tqsim_current_timestamp = self._api._account._get_current_timestamp()
             if tqsim_current_timestamp == 631123200000000000:
                 # 未收到任何行情, TqSim 时间没有更新
-                return tqsdk.TqApi._get_trading_day_start_time(self._data['_tqsdk_replay']['replay_dt'])
+                return _get_trading_day_start_time(self._data['_tqsdk_replay']['replay_dt'])
             else:
                 return tqsim_current_timestamp
         else:
